Move test_anaconda_token_with_private_package
